script_api/requirements.txt
requests==2.31.0
pandas==2.1.4
PyYAML==6.0.1
openpyxl==3.1.2